Correct migration with resource type id for resources of any project

+++ database/migrations/2020_09_29_182903_create_resource.php
@@ -16,7 +16,7 @@ class CreateResource extends Migration
         Schema::create('resources', function (Blueprint $table) {
 
             $table->id();
-			$table->unsignedSmallInteger('type_id');
+			$table->unsignedBigInteger('type_id');
 			$table->unsignedBigInteger('valuable_id');
 			$table->string('valuable_type');
